Reorganized ArcSight ESM API
